Created filtering file and moved filtertoolong there
